ci(infer): fix 2 pre-existing clippy -D warnings in cuda-feature path
Both only surface under the Mac typecheck combo
`--features cuda,no-cuda -- -D warnings` (the canonical no-nvcc pre-push
check); default and no-cuda feature sets were already clean. Pre-existing,
unrelated to any current feature work — they blocked the dependency compile
so infer's own lint state was invisible behind them.

- autograd/src/backend_cuda.rs:5690 — drop redundant `as u32` on
  `BLOCK_ONLINE / 32` (BLOCK_ONLINE is already u32; same-type cast). The
  divide result is bit-identical, zero behaviour change.
- cuda-kernels/src/kv_quant.rs:313 — reword a doc-comment continuation line
  that began with `+ ` (clippy parsed it as an unindented markdown list item);
  "rounded + clamped" -> "rounded then clamped". Comment-only.

Bench-exempt: one same-type-cast removal (provably bit-identical codegen) and
one doc-comment reword; neither is a runtime change. Verified clean with
`cargo clippy -p autograd` / `-p cuda-kernels --features cuda,no-cuda -- -D warnings`.

+++ crates/autograd/src/backend_cuda.rs
@@ -5687,7 +5687,7 @@ fn cuda_causal_sdpa_decode_gqa_cache(
     let use_online = head_dim == 256 && !env_force_legacy_decode_attn();
     if use_online {
         const BLOCK_ONLINE: u32 = 256; // = HEAD_DIM
-        let n_warps = (BLOCK_ONLINE / 32) as u32;
+        let n_warps = BLOCK_ONLINE / 32;
         let shared_online = n_warps * std::mem::size_of::<f32>() as u32;
         launch_rows(
             &backend.stream,
